experimental/OscarAPI/oscar_proj/requirements.txt
Babel==2.6.0
certifi==2018.11.29
chardet==3.0.4
dicttoxml==1.7.4
Django==2.1.4
django-appconf==1.0.2
django-compressor==2.2
django-extra-views==0.11.0
django-haystack==2.8.1
django-oscar==1.6.4
django-oscar-api==1.6.1
django-oscar-stores==1.1.1
django-phonenumber-field==2.0.1
django-tables2==1.21.2
django-treebeard==4.3
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
factory-boy==2.11.1
Faker==1.0.1
gunicorn==19.9.0
idna==2.7
lxml==4.2.5
mock==2.0.0
pbr==5.1.1
phonenumbers==8.10.2
Pillow==5.3.0
psycopg2-binary==2.8.3
purl==1.4
pycountry==18.12.8
python-dateutil==2.7.5
python-openpay==0.6.2
pytz==2018.7
rcssmin==1.0.6
requests==2.20.1
reverse==0.1.0
rjsmin==1.0.12
six==1.12.0
sorl-thumbnail==12.4.1
text-unidecode==1.2
Unidecode==1.0.23
urllib3==1.24.1
